OSN-594. Reorganize core data hook imports into separate files
