+++ ngui/ui/src/components/TopAlertWrapper/TopAlertWrapper.tsx
@@ -3,7 +3,7 @@ import { Box } from "@mui/material";
 import { render as renderGithubButton } from "github-buttons";
 import { FormattedMessage, useIntl } from "react-intl";
 import { useDispatch } from "react-redux";
-import { useAllDataSources } from "hooks/coreData";
+import { useAllDataSources } from "hooks/coreData/useAllDataSources";
 import { useGetToken } from "hooks/useGetToken";
 import { useOrganizationInfo } from "hooks/useOrganizationInfo";
 import { useRootData } from "hooks/useRootData";
