+++ ngui/ui/src/components/TopResourcesExpensesCard/TopResourcesExpensesCard.tsx
@@ -19,7 +19,7 @@ import TableLoader from "components/TableLoader";
 import TitleValue from "components/TitleValue";
 import Tooltip from "components/Tooltip";
 import WrapperCard from "components/WrapperCard";
-import { useOrganizationPerspectives } from "hooks/coreData";
+import { useOrganizationPerspectives } from "hooks/coreData/useOrganizationPerspectives";
 import { getLast30DaysResourcesUrl, getResourcesExpensesUrl, RESOURCE_PERSPECTIVES } from "urls";
 import { isEmpty as isEmptyArray } from "utils/arrays";
 import { FORMATTED_MONEY_TYPES } from "utils/constants";
