--- ngui/ui/src/hooks/coreData/index.ts
@@ -1,21 +1,0 @@
-import { useAllDataSources } from "./useAllDataSources";
-import { useCurrentEmployee } from "./useCurrentEmployee";
-import { useGetOptscaleCapability } from "./useGetOptscaleCapability";
-import { useInvitations } from "./useInvitations";
-import { useOrganizationAllowedActions } from "./useOrganizationAllowedActions";
-import { useOrganizationFeatures } from "./useOrganizationFeatures";
-import { useOrganizationPerspectives } from "./useOrganizationPerspectives";
-import { useOrganizations } from "./useOrganizations";
-import { useOrganizationThemeSettings } from "./useOrganizationThemeSettings";
-
-export {
-  useOrganizations,
-  useOrganizationAllowedActions,
-  useCurrentEmployee,
-  useAllDataSources,
-  useInvitations,
-  useOrganizationFeatures,
-  useGetOptscaleCapability,
-  useOrganizationThemeSettings,
-  useOrganizationPerspectives
-};
