+++ ngui/ui/src/hooks/useCustomOrganizationWeekends.ts
@@ -1,5 +1,5 @@
 import { SHORT_WEEK_DAYS } from "utils/datetime";
-import { useOrganizationFeatures } from "./coreData";
+import { useOrganizationFeatures } from "./coreData/useOrganizationFeatures";
 
 export const getValidCustomWeekends = (customWeekends) => {
   if (Array.isArray(customWeekends)) {
