+++ ngui/ui/src/hooks/useIsFeatureEnabled.ts
@@ -1,4 +1,4 @@
-import { useOrganizationFeatures } from "./coreData";
+import { useOrganizationFeatures } from "./coreData/useOrganizationFeatures";
 
 export const useIsFeatureEnabled = (featureName) => {
   const { [featureName]: featureFlag = 0 } = useOrganizationFeatures();
