+++ ngui/ui/src/hooks/useIsNebiusConnectionEnabled.ts
@@ -1,4 +1,4 @@
-import { useOrganizationFeatures } from "./coreData";
+import { useOrganizationFeatures } from "./coreData/useOrganizationFeatures";
 
 export const useIsNebiusConnectionEnabled = () => {
   const { nebius_connection_enabled: nebiusConnectionEnabled = 0 } = useOrganizationFeatures();
